small changes, added please wait label
The please wait label was added which is triggered on the click of a button on the initial scene, as loading the data can take some time. flag of chosen prices is set to true initially so that the standard selected prices (0 - 1000) are seen as valid, and so confirm is not needed unless prices are changed. This means getting to the main content of the program is much faster if you want to see the maximum range of property prices. Dates still need to be selected.